app/src/agent/pyproject.toml
[tool.poetry]
name = "legal-case-analyze-agent"
version = "0.1.0"
description = "A LangGraph-based legal case analysis agent using CopilotKit."
authors = ["Markus Ecker <markus.ecker@gmail.com>"]
license = "MIT"
packages = [
    { include = "legal_case_analyze_agent" }
]

[project]
name = "legal-case-analyze-agent"
version = "0.0.1"
dependencies = [
    "copilotkit[crewai]>=0.1.41",
    "langchain-openai>=0.2.3",
    "langchain-community>=0.3.1",
    "langchain-anthropic>=0.3.1",
    "langchain-google-genai>=2.0.5",
    "langchain>=0.3.4",
    "openai>=1.52.1",
    "tavily-python>=0.5.0",
    "python-dotenv>=1.0.1",
    "uvicorn[standard]>=0.34.0,<0.35.0",
    "requests>=2.32.3",
    "html2text>=2024.2.26",
    "langchain-core>=0.3.25",
    "langgraph-cli[inmem]>=0.1.64",
    "langgraph-checkpoint-sqlite>=2.0.1",
    "aiosqlite>=0.20.0",
    "aiohttp>=3.9.3",
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "fastapi (>=0.115.12,<0.116.0)"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.12.4,<3.13"
langchain = ">=0.3.4"
langgraph = "^0.0.28"
openai = "^1.14.2"
fastapi = ">=0.115.12,<0.116.0"
uvicorn = "^0.34.0"
python-dotenv = "^1.0.1"
copilotkit = {extras = ["crewai"], version = ">=0.1.41,<0.2.0"}

[tool.poetry.scripts]
demo = "research_canvas.demo:main"
